update and rename for configuration support
